# src/main/java/leoluiten/presentation/dtos/MatchDTO.java
package leoluiten.presentation.dtos;

import jakarta.validation.constraints.NotNull;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class MatchDTO {

    @NotNull
    private Long gameId;

    @NotNull
    private Long playerId;
}